requirements.txt
pillow
cryptography
requests